requirements.txt
ansible==..
ansible-compat==..
ansible-core==..
ansible-lint==..
attrs==..
beautifulsoup==..
black==..
blinker==..
bracex==..post
certifi==..
cffi==..
charset-normalizer==..
click==..
cryptography==..
filelock==..
Flask==..
idna==.
importlib_metadata==..
itsdangerous==..
Jinja==..
jmespath==..
jsonschema==..
jsonschema-specifications==..
lxml==..
MarkupSafe==..
mypy_extensions==..
packaging==.
pathspec==..
platformdirs==..
pycparser==.
PyYAML==..
referencing==..
requests==..
resolvelib==..
rpds-py==..
ruamel.yaml==..
ruamel.yaml.clib==..
soupsieve==.
subprocess-tee==..
typing_extensions==..
urllib==..
wcmatch==.
Werkzeug==..
yamllint==..
zipp==..
